refactor(validators): partial cleanup of legacy validator architecture
- Removed obsolete validators (argument, attribute, function, module, system, python)
- Began replacing static validation flow with context-aware logic
- Introduced early wiring for runtime/module validation in Spy entrypoint
- Stubbed modular validator orchestration using dedicated runtime/module classes
- Temporary API degradation in function/module validation pending reimplementation

Note: This commit breaks some validation logic. Work is paused mid-refactor.
